macOS 10.15 beta 2, Xcode 11 beta 2 (iOS, tvOS, watchOS, iOSSupport 13)

+++ macOS/PrivateFrameworks/GeoServices.framework/GEOAnalyticsPipelineStateFactory.h
@@ -22,7 +22,7 @@ __attribute__((visibility("hidden")))
 + (id)sharedFactory;
 - (void).cxx_destruct;
 - (void)deviceCountryChanged:(id)arg1;
-- (id)sessionStateForType:(int)arg1;
+- (id)sessionStateForType:(long long)arg1;
 - (id)_emptyUserSessionState;
 - (id)stateForType:(long long)arg1;
 - (id)placeRequest;
